python/requirements.txt
faster-whisper>=1.0.0
llama-cpp-python>=0.2.0
piper-tts>=1.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
